tests/wpt/tests/html/dom/render-blocking/element-render-blocking-039.html

<!DOCTYPE html>
<head>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="support/utils.js"></script>
<title>link rel=expect: render blocking stops waiting when the document finishes loading</title>

<link rel=expect href="#invalid" blocking="render">
<script>
const startTime = performance.now();

async_test((t) => {
  requestAnimationFrame(() => {
    t.step(() => assert_less_than(performance.now() - startTime, 3000, "requestAnimationFrame should not have been delayed by the link that didn't find its target after load completes"));
    t.done();
  });
}, "");
</script>
</head>
<body>
  <div id="first"></div>
  <div id="last"></div>
</body>
